requirements.txt
PyQt5==5.15.9
mysql-connector-python==8.1.0
matplotlib==3.7.2
numpy==1.24.3
python-dotenv==1.0.0
